calcCart 관련 함수는 별도 파일로 분리. 데이터, 액션, 계산으로 분리 후 전체 테스트 통과
