+++ src/constants.js
@@ -1,0 +1,8 @@
+export const ElementIds = {
+  SUM: "cart-total",
+  SEL: "product-select",
+  ADD_BTN: "add-to-cart",
+  STOCK_INFO: "stock-status",
+  LOYALTY_POINTS: "loyalty-points",
+  CART_DISP: "cart-items",
+};
